+++ src/main/java/com/testcar/car/domains/car/entity/Car.java
@@ -1,4 +1,4 @@
-package com.testcar.car.domains.car;
+package com.testcar.car.domains.car.entity;
 
 
 import com.testcar.car.common.entity.BaseEntity;
