+++ src/main/java/com/testcar/car/domains/car/model/CarResponse.java
@@ -2,8 +2,8 @@ package com.testcar.car.domains.car.model;
 
 
 import com.testcar.car.common.annotation.DateTimeFormat;
-import com.testcar.car.domains.car.Car;
-import com.testcar.car.domains.car.Type;
+import com.testcar.car.domains.car.entity.Car;
+import com.testcar.car.domains.car.entity.Type;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 import lombok.Builder;
