+++ src/main/java/com/testcar/car/domains/car/model/vo/CarFilterCondition.java
@@ -2,7 +2,7 @@ package com.testcar.car.domains.car.model.vo;
 
 
 import com.testcar.car.common.annotation.DateTimeFormat;
-import com.testcar.car.domains.car.Type;
+import com.testcar.car.domains.car.entity.Type;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 import lombok.Getter;
