+++ src/main/java/com/testcar/car/domains/carReservation/entity/CarReservation.java
@@ -1,8 +1,8 @@
-package com.testcar.car.domains.carReservation;
+package com.testcar.car.domains.carReservation.entity;
 
 
 import com.testcar.car.common.entity.BaseEntity;
-import com.testcar.car.domains.carStock.CarStock;
+import com.testcar.car.domains.carStock.entity.CarStock;
 import com.testcar.car.domains.member.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
