+++ src/main/java/com/testcar/car/domains/carStock/repository/CarStockCustomRepository.java
@@ -1,7 +1,7 @@
 package com.testcar.car.domains.carStock.repository;
 
 
-import com.testcar.car.domains.carStock.CarStock;
+import com.testcar.car.domains.carStock.entity.CarStock;
 import com.testcar.car.domains.carStock.model.vo.CarStockFilterCondition;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
