+++ src/main/java/com/testcar/car/domains/carStock/repository/CarStockCustomRepositoryImpl.java
@@ -7,8 +7,8 @@ import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.testcar.car.common.entity.BaseQueryDslRepository;
-import com.testcar.car.domains.carStock.CarStock;
-import com.testcar.car.domains.carStock.StockStatus;
+import com.testcar.car.domains.carStock.entity.CarStock;
+import com.testcar.car.domains.carStock.entity.StockStatus;
 import com.testcar.car.domains.carStock.model.vo.CarStockFilterCondition;
 import java.util.List;
 import java.util.Optional;
