+++ src/main/java/com/testcar/car/domains/carStock/repository/CarStockRepository.java
@@ -1,7 +1,7 @@
 package com.testcar.car.domains.carStock.repository;
 
 
-import com.testcar.car.domains.carStock.CarStock;
+import com.testcar.car.domains.carStock.entity.CarStock;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
